Fix strictnulls.

+++ app/services/scene-collections/scene-collections.ts
@@ -1111,7 +1111,7 @@ export class SceneCollectionsService extends Service implements ISceneCollection
    * @param sceneId - The scene id
    */
   removeNodeMap(sceneId: string) {
-    if (!this.sceneNodeMaps[sceneId]) return;
+    if (!this.sceneNodeMaps || this.sceneNodeMaps[sceneId]) return;
 
     this.stateService.removeNodeMap(sceneId);
   }
